feature/order/save-new-order: completed save a new order when user ordered by cash and VNPay payment case

+++ shoes-store-backend/payment-service/src/main/java/ktpm17ctt/g6/paymentservice/dtos/responses/PaymentResponse.java
@@ -10,7 +10,7 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class PaymentResponse {
     String orderId;
-    double amount;
+    long amount;
     String status;
     String transactionId;
     String paymentUrl;
